cleaned up build scripts

+++ examples/lambda-calculus/build.bat
@@ -259,19 +259,19 @@ if not exist "%_BUILD_DIR%\build" (
     )
     call "%_JAVA_CMD%" -jar "%_FLIX_JAR%" init
 )
+if exist "%_BUILD_DIR%\src\*.flix" del /q "%_BUILD_DIR%\src\*.flix"
+if exist "%_BUILD_DIR%\test\*.flix" del /q "%_BUILD_DIR%\test\*.flix"
+
 @rem xcopy must be called AFTER flix init
 if %_DEBUG%==1 ( echo %_DEBUG_LABEL% xcopy /s /y "%_SOURCE_MAIN_DIR%" "%_BUILD_DIR%\src\" 1^>NUL 1>&2
 ) else if %_VERBOSE%==1 ( echo Copy %__N_FILES% to directory "!_BUILD_DIR:%_ROOT_DIR%=!\src\" 1>&2
 )
-xcopy /s /y "%_SOURCE_MAIN_DIR%" "%_BUILD_DIR%\src\" to directory "!_BUILD_DIR:%_ROOT_DIR%=!\src\" 1>NUL
+xcopy /s /y "%_SOURCE_MAIN_DIR%" "%_BUILD_DIR%\src\" 1>NUL
 if not %ERRORLEVEL%==0 (
     popd
-    echo %_ERROR_LABEL% Failed to copy %__N_FILES% 1>&2
+    echo %_ERROR_LABEL% Failed to copy %__N_FILES% to directory "!_BUILD_DIR:%_ROOT_DIR%=!\src\" 1>&2
     set _EXITCODE=1
     goto :eof
-)
-if exist "%_BUILD_DIR%\test\*.flix" (
-    del /q "%_BUILD_DIR%\test\*.flix"
 )
 call :compile_lib
 if not %_EXITCODE%==0 popd & goto :eof
